+++ str.h
@@ -1,10 +1,12 @@
 #ifndef STR_H
 #define STR_H
 
-extern unsigned int str_copy(char *,char *);
+#include <sys/types.h>
+
+extern off_t str_copy(char *,char *);
 extern int str_diff(char *,char *);
-extern int str_diffn(char *,char *,unsigned int);
-extern unsigned int str_len(char *);
+extern int str_diffn(char *,char *,off_t);
+extern off_t str_len(char *);
 extern unsigned int str_chr(char *,int);
 extern unsigned int str_rchr(char *,int);
 extern int str_start(char *,char *);
